+++ internal/schema/validate.go
@@ -20,6 +20,7 @@ var (
 		"scope_name":      ScopeName,
 		"collection_name": CollectionName,
 		"index_name":      IndexName,
+		"keyspace":        Keyspace,
 		"app_service_id":  AppServiceId,
 		"endpoint_id":     EndpointId,
 	}
